# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: MiroFish-Novel API v2 Baseline
  version: 0.1.0
  description: |
    Rebuild contract baseline for operation, error, and health resources.
servers:
  - url: http://127.0.0.1:5102
...
paths:
...
  /api/v2/workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces
      parameters:
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Query one workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Workspace detail
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/workspaces/{workspaceId}/ingests:
    post:
      operationId: createWorkspaceIngest
      tags: [workspaces]
      summary: Create one ingest workflow
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Ingest operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
...
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Stable error payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ApiError:
      type: object
      required: [code, message, details, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        retryable:
          type: boolean
    ErrorResponse:
      type: object
      required: [trace_id, error]
      properties:
        trace_id:
          type: string
        error:
          $ref: "#/components/schemas/ApiError"
    OperationStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    OperationStepStatus:
      type: string
      enum: [pending, running, succeeded, failed, skipped]
    OperationStep:
      type: object
      required: [step_id, stage, status, progress_percent]
      properties:
        step_id:
          type: string
        stage:
          type: string
        status:
          $ref: "#/components/schemas/OperationStepStatus"
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        elapsed_ms:
          type: integer
          nullable: true
...
    OperationResource:
      type: object
      required:
        - operation_id
        - trace_id
        - status
        - progress_percent
        - retryable
        - steps
      properties:
        operation_id:
          type: string
        trace_id:
          type: string
        status:
          $ref: "#/components/schemas/OperationStatus"
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        retryable:
          type: boolean
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/OperationStep"